+++ theme/diverse/lib.php
@@ -42,8 +42,10 @@ function theme_diverse_get_main_scss_content($theme) {
     // This way, DIVERSE will ship the same SCSS code as Boost Union itself.
     $scss = theme_boost_union_get_main_scss_content(\core\output\theme_config::load('boost_union'));
 
-    // And add DIVERSE's main SCSS file to the stack.
-    $scss .= file_get_contents($CFG->dirroot . '/theme/diverse/scss/post.scss');
+    // And add DIVERSE's main SCSS file and the page-specific SCSS files to the stack.
+    foreach (['post', 'landing', 'login'] as $file) {
+        $scss .= file_get_contents($CFG->dirroot . '/theme/diverse/scss/' . $file . '.scss');
+    }
 
     return $scss;
 }